+++ autocnet/matcher/subpixel.py
@@ -1,45 +1,37 @@
 from collections import defaultdict
+from contextlib import nullcontext
 import json
 import logging
 from math import floor
 import numbers
-import sys
 import time
 import warnings
 import logging
 
 import numpy as np
-
-from scipy.ndimage.measurements import center_of_mass
+import pvl
+from scipy.ndimage import center_of_mass
 from skimage import transform as tf
 from skimage import registration
 from skimage import filters
 from scipy import fftpack
 from scipy.spatial import distance_matrix
-import numpy as np
 from sqlalchemy.sql.expression import bindparam
-
+from sqlalchemy import inspect
 from matplotlib import pyplot as plt
-
-from plio.io.io_gdal import GeoDataset
-
-import pvl
-
-import PIL
 
 from autocnet.matcher.naive_template import pattern_match
 from autocnet.matcher.mutual_information import mutual_information
-from autocnet.spatial import isis
+from autocnet.io import isis
+from autocnet.io.geodataset import AGeoDataset
 from autocnet.io.db.model import Measures, Points, Images, JsonEncoder
 from autocnet.graph.node import NetworkNode
 from autocnet.transformation import roi
 from autocnet.transformation.affine import estimate_local_affine
-from autocnet import spatial
 from autocnet.utils.serializers import JsonEncoder
 
-from sqlalchemy import inspect
 
-PIL.Image.MAX_IMAGE_PIXELS = sys.float_info.max
+
 log = logging.getLogger(__name__)
 
 
@@ -368,7 +360,10 @@ def subpixel_register_point(pointid,
         nodes = {}
         for measure in measures:
             res = session.query(Images).filter(Images.id == measure.imageid).one()
-            nodes[measure.imageid] = NetworkNode(node_id=measure.imageid, image_path=res.path)
+            nodes[measure.imageid] = NetworkNode(node_id=measure.imageid, 
+                                                 image_path=res.path, 
+                                                 dem=res.dem,
+                                                 dem_type=res.dem_type)
         session.expunge_all()
 
     resultlog = []
@@ -537,7 +532,7 @@ def register_to_base(pointid,
     session = ncg.Session()
 
     if isinstance(base_image, str):
-        base_image = GeoDataset(base_image)
+        base_image = AGeoDataset(base_image)
 
     if isinstance(pointid, Points):
         point = pointid
@@ -554,13 +549,11 @@ def register_to_base(pointid,
         measures = point.measures
 
         # Attempt to project the point into the base image
-        bpoint = spatial.isis.point_info(base_image.file_name, point.geom.x, point.geom.y, 'ground')
-        if bpoint is None:
+        bsample, bline = base_image.sensormodel.lonlat2sampline(point.geom.x, point.geom.y)
+        if bsample is None:
             log.warning('unable to find point in ground image')
             # Need to set the point to False
             return
-        bline = bpoint.get('Line')
-        bsample = bpoint.get('Sample')
 
         # Setup a cache so that we can get the file handles one time instead of
         # once per measure in the measures list.
@@ -576,7 +569,7 @@ def register_to_base(pointid,
             try:
                 measure_image = image_cache[res.id]
             except:
-                measure_image = GeoDataset(measure_image.path)
+                measure_image = AGeoDataset(measure_image.path)
                 image_cache[res.id] = measure_image
 
             # Attempt to match the base
@@ -829,11 +822,12 @@ def fourier_mellen(ref_image, moving_image, affine=tf.AffineTransform(), verbose
     return subpixel_affine, error, diffphase
 
 def subpixel_register_point_smart(point,
-                                  session,
+                                  session=None,
                                   cost_func=lambda x,y: 1/x**2 * y,
                                   parameters=[],
                                   chooser='subpixel_register_point_smart',
-                                  verbose=False):
+                                  verbose=False,
+                                  ncg=None):
 
     """
     Given some point, subpixel register all of the measures in the point to the
@@ -863,38 +857,35 @@ def subpixel_register_point_smart(point,
                  {'match_kwargs': {'image_size':(151,151), 'template_size':(67,67)}},
                  {'match_kwargs': {'image_size':(181,181), 'template_size':(73,73)}}]
     """
+    with ncg.session_scope() if ncg is not None else nullcontext(session) as session:
+        if not isinstance(point, Points):
+            point = session.query(Points).filter(Points.id == point).one()
+        pointid = point.id
 
+        # Order by is important here because the measures get ids in sequential order when initially placed
+        # and the reference_index is positionally linked to the ordered vector of measures.
+        measures = session.query(Measures).filter(Measures.pointid == pointid).order_by(Measures.id).all()
+        reference_index = point.reference_index
 
-    # if not ncg.Session:
-    #     raise BrokenPipeError('This func requires a database session from a NetworkCandidateGraph.')
+        # Get the reference measure to instantiate the source node. All other measures will
+        # match to the source node.
+        source = measures[reference_index]
+        reference_index_id = source.imageid
+        log.info(f'Using measure {source.id} on image {source.imageid}/{source.serial} as the reference.')
+        log.info(f'Measure reference index is: {reference_index}')
 
-    if not isinstance(point, Points):
-        point = session.query(Points).filter(Points.id == point).one()
-    pointid = point.id
-
-    #with ncg.session_scope() as session:
-    # Order by is important here because the measures get ids in sequential order when initially placed
-    # and the reference_index is positionally linked to the ordered vector of measures.
-    measures = session.query(Measures).filter(Measures.pointid == pointid).order_by(Measures.id).all()
-    reference_index = point.reference_index
-
-    # Get the reference measure to instantiate the source node. All other measures will
-    # match to the source node.
-    source = measures[reference_index]
-    reference_index_id = source.imageid
-    
-    log.info(f'Using measure {source.id} on image {source.imageid}/{source.serial} as the reference.')
-    log.info(f'Measure reference index is: {reference_index}')
-
-    # Build a node cache so that this is an encapsulated database call. Then nodes
-    # can be pulled from the lookup sans database.
-    nodes = {}
-    for measure in measures:
-        res = session.query(Images).filter(Images.id == measure.imageid).one()
-        nn = NetworkNode(node_id=measure.imageid, image_path=res.path)
-        #nn.parent = ncg
-        nodes[measure.imageid] = nn
-        session.expunge_all()
+        # Build a node cache so that this is an encapsulated database call. Then nodes
+        # can be pulled from the lookup sans database.
+        nodes = {}
+        for measure in measures:
+            res = session.query(Images).filter(Images.id == measure.imageid).one()
+            nn = NetworkNode(node_id=measure.imageid, 
+                            image_path=res.path,
+                            cam_type=res.cam_type,
+                            dem=res.dem,
+                            dem_type=res.dem_type)
+            nodes[measure.imageid] = nn
+            session.expunge_all()
 
     log.info(f'Attempting to subpixel register {len(measures)-1} measures for point {pointid}')
     # Set the reference image
@@ -1038,7 +1029,7 @@ def subpixel_register_point_smart(point,
                 new_x, new_y = updated_affine([measure.apriorisample, measure.aprioriline])[0]
                 
                 dist = np.linalg.norm([measure.aprioriline-new_x, 
-                                      measure.apriorisample-new_y])
+                                       measure.apriorisample-new_y])
                 cost = cost_func(dist, metric)
 
                 m = {'id': measure.id,
@@ -1177,7 +1168,7 @@ def decider(measures, tol=0.5):
     return measures_to_update, measures_to_set_false
 
 def validate_candidate_measure(measure_to_register,
-                               session,
+                               session=None,
                                ncg=None,
                                parameters=[],
                                **kwargs):
@@ -1219,28 +1210,30 @@ def validate_candidate_measure(measure_to_register,
 
     measure_to_register_id = measure_to_register['id']
 
-    # Get the measure to be registered
-    measure = session.query(Measures).filter(Measures.id == measure_to_register_id).order_by(Measures.id).one()
+    with ncg.session_scope() if ncg is not None else nullcontext(session) as session:
+        # Get the measure to be registered
+        measure = session.query(Measures).filter(Measures.id == measure_to_register_id).order_by(Measures.id).one()
 
-    # Get the references measure
-    point = measure.point
-    reference_index = point.reference_index
-    reference_measure = point.measures[reference_index]
+        # Get the references measure
+        point = measure.point
+        reference_index = point.reference_index
+        reference_measure = point.measures[reference_index]
 
 
-    # Match the reference measure to the measure_to_register - this is the inverse of the first match attempt
-    # Source is the image that we are seeking to validate, destination is the reference measure.
-    # This is the inverse of other functions as this is a validator.
+        # Match the reference measure to the measure_to_register - this is the inverse of the first match attempt
+        # Source is the image that we are seeking to validate, destination is the reference measure.
+        # This is the inverse of other functions as this is a validator.
 
-    source_imageid = measure.imageid
-    source_image = session.query(Images).filter(Images.id == source_imageid).one()
-    source_node = NetworkNode(node_id=source_imageid, image_path=source_image.path)
-    source_node.parent = ncg
+        source_imageid = measure.imageid
+        source_image = session.query(Images).filter(Images.id == source_imageid).one()
+        source_node = NetworkNode(node_id=source_imageid, image_path=source_image.path)
+        source_node.parent = ncg
 
-    destination_imageid = reference_measure.imageid
-    destination_image = session.query(Images).filter(Images.id == destination_imageid).one()
-    destination_node = NetworkNode(node_id=destination_imageid, image_path=destination_image.path)
-    destination_node.parent = ncg
+        destination_imageid = reference_measure.imageid
+        destination_image = session.query(Images).filter(Images.id == destination_imageid).one()
+        destination_node = NetworkNode(node_id=destination_imageid, image_path=destination_image.path)
+        destination_node.parent = ncg
+        session.expunge_all()
 
     sample = measure_to_register['sample']
     line = measure_to_register['line']
@@ -1252,13 +1245,13 @@ def validate_candidate_measure(measure_to_register,
                             line, 
                             size_x=parameters[0]['match_kwargs']['image_size'][0],
                             size_y=parameters[0]['match_kwargs']['image_size'][1], 
-                            buffer=10)
+                            buffer=0)
     moving_roi = roi.Roi(destination_node.geodata, 
                             reference_measure.sample, 
                             reference_measure.line, 
                             size_x=parameters[0]['match_kwargs']['template_size'][0],
                             size_y=parameters[0]['match_kwargs']['template_size'][1],
-                            buffer=10)
+                            buffer=20)
 
     try:
         baseline_affine = estimate_local_affine(reference_roi, moving_roi)
@@ -1276,13 +1269,13 @@ def validate_candidate_measure(measure_to_register,
                                 line, 
                                 size_x=match_kwargs['image_size'][0],
                                 size_y=match_kwargs['image_size'][1], 
-                                buffer=10)
+                                buffer=0)
         moving_roi = roi.Roi(destination_node.geodata, 
                                 reference_measure.sample, 
                                 reference_measure.line, 
                                 size_x=match_kwargs['template_size'][0],
                                 size_y=match_kwargs['template_size'][1],
-                                buffer=10)
+                                buffer=20)
 
         # Handle the exception where the clip can raise an index error if it is outside the image
         try:
@@ -1305,12 +1298,11 @@ def validate_candidate_measure(measure_to_register,
     return dists
 
 def smart_register_point(point, 
-                         session,
+                         session=None,
                          parameters=[], 
                          shared_kwargs={}, 
                          valid_reprojection_distance=1.1, 
-                         ncg=None, 
-                         Session=None):
+                         ncg=None):
     """
     The entry func for the smart subpixel registration code. This is the user
     side API func for subpixel registering a point using the smart matcher.
@@ -1356,9 +1348,11 @@ def smart_register_point(point,
                             building approach
 
     """
-    if not isinstance(point, Points):
-        point = session.query(Points).filter(Points.id == point).one()
-
+    with ncg.session_scope() if ncg is not None else nullcontext(session) as session:
+        if not isinstance(point, Points):
+            point = session.query(Points).filter(Points.id == point).one()
+        session.expunge_all()
+        
     measure_results = subpixel_register_point_smart(point, session, parameters=parameters, **shared_kwargs)
     measures_to_update, measures_to_set_false = decider(measure_results)
     log.info(f'Found {len(measures_to_update)} measures that found subpixel registration consensus. Running validation now...')